+++ lib/common/color.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 
-class EpicTrackerColors {
-  EpicTrackerColors._();
-
+abstract class EpicTrackerColors {
   static const Color main = Color(0xff222831);
-  static const Color secondary = Color(0xff00ADB5);
+  static const Color secondary = Color(0xff994FF3);
   static const Color accent = Color(0xffEEEEEE);
   static const Color accentBlack = Color(0xff393E46);
   static const Color intensePink = Color(0xffD9376E);