+++ lib/config/config.dart
@@ -1,0 +1,2 @@
+export 'package:epicticker/config/routes/main_routes.dart';
+export 'package:epicticker/config/theme/app_theme.dart';