+++ lib/config/theme/app_theme.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
 
-class AppTheme {
-  AppTheme._();
-
+abstract class AppTheme {
   static ThemeData getTheme() => ThemeData(
     useMaterial3: false,
-    brightness: Brightness.dark
+    brightness: Brightness.dark,
   );
 }