+++ lib/presentation/screens/home_screen.dart
@@ -2,11 +2,14 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:epicticker/common/color.dart';
 import 'package:epicticker/domain/entities/count_down_entity.dart';
-import 'package:epicticker/presentation/providers/crud/count_down_provider.dart';
+import 'package:epicticker/presentation/providers/count_down_provider.dart';
 import 'package:epicticker/presentation/screens/edit_countdown_screen.dart';
 import 'package:epicticker/presentation/widgets/animation_limited_widget.dart';
-import 'package:epicticker/presentation/widgets/dashboard_widget.dart';
-import 'package:epicticker/presentation/widgets/day_left_card_widget.dart';
+import 'package:epicticker/presentation/widgets/dashboard/dashboard_widget.dart';
+import 'package:epicticker/presentation/widgets/day_left_card/day_left_card_widget.dart';
+import 'package:epicticker/utils/difference_date_util.dart';
+import 'package:epicticker/utils/get_month_name.dart';
+import 'package:epicticker/utils/time_remaining_format.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
@@ -20,7 +23,7 @@ class _HomeScreenState extends State<HomeScreen> {
   @override
   void initState() {
     super.initState();
-    Provider.of<CountDownProvider>(context, listen: false).loadCountDowns();
+    Provider.of<CountDownProvider>(context, listen: false).getAllCountdowns();
   }
 
   Widget bodyContainer(BuildContext context) {
@@ -29,45 +32,10 @@ class _HomeScreenState extends State<HomeScreen> {
     	child: Column(
     		children: <Widget>[
     			const DashboardWidget(),
-    			Consumer<CountDownProvider>(
-    				builder: (BuildContext context, CountDownProvider countdownProvider, Widget? child) {
-    					final List<CountDownEntity> countdowns = countdownProvider.countDownList;
-
-    					if (countdowns.isEmpty) const Center(child: Text('No countdowns yet.'));
-
-    					return Container(
-    						margin: const EdgeInsets.only(bottom: 100.0),
-    						child: Column(
-    							children: countdowns.asMap().entries.map((MapEntry<int, CountDownEntity> entry) {
-    								final int index = entry.key;
-    								final CountDownEntity countdown = entry.value;
-
-    								return AnimationLimiterWidget(
-    									position: index,
-    									child: InkWell(
-    										onTap: () {
-    											CountDownEntity currentCountdown = countdown;
-
-    											Navigator.push(
-    												context,
-    												MaterialPageRoute<CountDownEntity>(
-    													builder: (_) => EditCountDownScreen(currentCountdown: currentCountdown)
-    												)
-    											);
-    										},
-    										child: DayLeftCardWidget(
-    											title: countdown.name,
-    											year: countdown.year,
-    											month: countdown.month,
-    											day: countdown.day
-    										)
-    									),
-    								);
-    							}).toList(),
-    						),
-    					);
-    				}
-    			),
+    			Container(
+						margin: const EdgeInsets.symmetric(vertical: 20.0),
+						child: _CardList()
+					),
     		],
     	),
     );
@@ -82,9 +50,89 @@ class _HomeScreenState extends State<HomeScreen> {
 					behavior: const ScrollBehavior().copyWith(overscroll: true),
 					child: SingleChildScrollView(child: bodyContainer(context))
 				)
-			),
-			// floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
-			// floatingActionButton: const FloatingButtonWidget()
+			)
+    );
+  }
+}
+
+class _CardList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Consumer<CountDownProvider>(
+    	builder: (BuildContext context, CountDownProvider countdownProvider, Widget? child) {
+    		final List<CountdownEntity> countdowns = countdownProvider.getRecentCountdownByDays();
+
+    		if (countdowns.isEmpty) const Center(child: Text('No countdowns yet.'));
+
+    		return Container(
+    			margin: const EdgeInsets.only(bottom: 100.0),
+    			child: Column(
+    				children: countdowns.asMap().entries.map((MapEntry<int, CountdownEntity> entry) {
+    					final int index = entry.key;
+    					final CountdownEntity countdown = entry.value;
+
+							DateDifference difference = calculateDateDifference(
+								countdown.year,
+								countdown.month,
+								countdown.day,
+							);
+
+							double completionPercentage =
+								1.0 - (difference.years * 365 + difference.months * 30 + difference.days) / 365;
+
+							completionPercentage = completionPercentage.clamp(0.0, 1.0);
+
+    					return _Card(
+								index: index,
+								countdown: countdown,
+								difference: difference,
+								completionPercentage: completionPercentage
+							);
+    				}).toList(),
+    			),
+    		);
+    	}
+    );
+  }
+}
+
+class _Card extends StatelessWidget {
+  const _Card({
+    required this.index,
+    required this.countdown,
+    required this.difference,
+    required this.completionPercentage,
+  });
+
+  final int index;
+  final CountdownEntity countdown;
+  final DateDifference difference;
+  final double completionPercentage;
+
+  @override
+  Widget build(BuildContext context) {
+    return AnimationLimiterWidget(
+    	position: index,
+    	child: InkWell(
+    		onTap: () {
+    			CountdownEntity currentCountdown = countdown;
+
+					// TODO: change this for go router
+    			Navigator.push(
+    				context,
+    				MaterialPageRoute<CountdownEntity>(
+    					builder: (_) => EditCountDownScreen(currentCountdown: currentCountdown)
+    				)
+    			);
+    		},
+    		child: DayLeftCardWidget(
+					// TODO: change this for title
+					title: countdown.name,
+					daysLeft: timeRemainingFormat(difference.years, difference.months, difference.days),
+					targetDate: '${countdown.day} ${SelectMonth.getCompletedMonthName(countdown.month)}, ${countdown.year}',
+					completion: completionPercentage,
+				)
+    	),
     );
   }
 }