+++ lib/utils/convert_string_filter_chip.dart
@@ -1,0 +1,16 @@
+import 'package:epicticker/domain/entities/count_down_entity.dart';
+
+FilterChipType? convertStringToFilterChipType(String? filterTypeString) {
+  if (filterTypeString == null) return null;
+
+  switch (filterTypeString) {
+    case 'activated':
+      return FilterChipType.activated;
+    case 'archived':
+      return FilterChipType.archived;
+    case 'completed':
+      return FilterChipType.completed;
+    default:
+      return null;
+  }
+}