+++ lib/utils/get_month_name.dart
@@ -1,18 +1,52 @@
-String getMonthName(int month) {
-  final List<String> months = <String>[
-    'January',
-    'February',
-    'March',
-    'April',
-    'May',
-    'June',
-    'July',
-    'August',
-    'September',
-    'October',
-    'November',
-    'December'
-  ];
+abstract class SelectMonth {
+	static String getCompletedMonthName(int month) {
+    final List<String> months = <String>[
+      'January',
+      'February',
+      'March',
+      'April',
+      'May',
+      'June',
+      'July',
+      'August',
+      'September',
+      'October',
+      'November',
+      'December'
+    ];
 
-  return months[month - 1];
+    return months[month - 1];
+  }
+
+	static String getShortMonthName(int month) {
+		final List<String> months = <String>[
+      'Jan',
+      'Feb',
+      'Mar',
+      'Apr',
+      'May',
+      'Jun',
+      'Jul',
+      'Aug',
+      'Sep',
+      'Oct',
+      'Nov',
+      'Dec'
+    ];
+
+    return months[month - 1];
+	}
+
+	static String getWeekName(int day) {
+    final List<String> daysOfWeek = <String>[
+      'Monday',
+      'Tuesday',
+      'Wednesday',
+      'Thursday',
+      'Friday',
+      'Saturday',
+      'Sunday'
+    ];
+    return daysOfWeek[day - 1];
+  }
 }