+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Added
 
+- New Card of Day Left Added | [#58](https://github.com/reliek21/epicticker-mobile-app/pull/58) | [reliek21](https://github.com/reliek21)
+- Prepare for the new design | [#45](https://github.com/reliek21/epicticker-mobile-app/pull/45) | [reliek21](https://github.com/reliek21)
 - App Icon and Name changed | [#29](https://github.com/reliek21/epicticker-mobile-app/pull/29) | [reliek21](https://github.com/reliek21)
 - New Dashboard design | [#29](https://github.com/reliek21/epicticker-mobile-app/pull/29) | [reliek21](https://github.com/reliek21)
 - Validations for the CRUD | [#24](https://github.com/reliek21/epicticker-mobile-app/pull/24) | [reliek21](https://github.com/reliek21)
@@ -18,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Updated
 
 ### Refactor
+- New Design of the dashboard added | [#59](https://github.com/reliek21/epicticker-mobile-app/pull/59) | [reliek21](https://github.com/reliek21)
 - Refactor of code in the app | [#15](https://github.com/reliek21/epicticker-mobile-app/pull/15) | [reliek21](https://github.com/reliek21)
 - Refactor of components of Dashboard and utils | [#11](https://github.com/reliek21/epicticker-mobile-app/pull/11) | [reliek21](https://github.com/reliek21)
 - Linter anotation added | [#4](https://github.com/reliek21/epicticker-mobile-app/pull/4) | [reliek21](https://github.com/reliek21)
@@ -28,5 +31,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ### Dependencies
 
+- build(deps): bump provider from 6.1.1 to 6.1.2 | [#67](https://github.com/reliek21/epicticker-mobile-app/pull/67) | [reliek21](https://github.com/reliek21)
+- build(deps): bump intl from 0.18.1 to 0.19.0 | [#61](https://github.com/reliek21/epicticker-mobile-app/pull/61) | [reliek21](https://github.com/reliek21)
+- build(deps): bump go_router from 12.1.1 to 13.0.0 | [#62](https://github.com/reliek21/epicticker-mobile-app/pull/62) | [reliek21](https://github.com/reliek21)
 - build(deps): bump flutter_lints from 3.0.0 to 3.0.1 | [#28](https://github.com/reliek21/epicticker-mobile-app/pull/28) | [reliek21](https://github.com/reliek21)
-
+- build(deps): bump provider from 6.0.5 to 6.1.1 [#44](https://github.com/reliek21/epicticker-mobile-app/pull/44) | [reliek21](https://github.com/reliek21)